py-service/src/requirements.txt
mysql-connector-python==8.0.29
Flask==2.1.2
